fix: resolve all remaining compilation warnings

+++ Minecraft.Client/Rendering/LevelRenderer.cpp
@@ -2390,7 +2390,7 @@ void LevelRenderer::setDirty(int x0, int y0, int z0, int x1, int y1, int z1, Lev
 
 					dirtyChunksLockFreeStack.Push((int *)(index));
 #else
-					dirtyChunksLockFreeStack.Push((int *)(uintptr_t)(index + 2));		
+					dirtyChunksLockFreeStack.Push((int *)(intptr_t)(uintptr_t)(index + 2));		
 #endif
 
 #ifdef _XBOX
@@ -3603,13 +3603,14 @@ void LevelRenderer::staticCtor()
 {
 	s_rebuildCompleteEvents = new C4JThread::EventArray(MAX_CHUNK_REBUILD_THREADS);
 	char threadName[256];
-	        for(unsigned int i = 0; i < MAX_CHUNK_REBUILD_THREADS; ++i)
-	        {
-	                sprintf(threadName,"Rebuild Chunk Thread %d\n",i);
-	                rebuildThreads[i] = new C4JThread(rebuildChunkThreadProc,(void *)(uintptr_t)i,threadName);
-	
-	                s_activationEventA[i] = new C4JThread::Event();
-			// Threads 1,3 and 5 are generally idle so use them
+	for(unsigned int i = 0; i < MAX_CHUNK_REBUILD_THREADS; ++i)
+	{
+		sprintf(threadName,"Rebuild Chunk Thread %d\n",i);
+		rebuildThreads[i] = new C4JThread(rebuildChunkThreadProc,(void *)i,threadName);
+
+		s_activationEventA[i] = new C4JThread::Event();
+
+		// Threads 1,3 and 5 are generally idle so use them
 		if((i%3) == 0) rebuildThreads[i]->SetProcessor(CPU_CORE_CHUNK_REBUILD_A);
 		else if((i%3) == 1)
 		{
